kindle_reviews/requirements.txt
pyspark==3.1.2
nltk==3.4
pyyaml==5.4.1
setuptools==58.0.4